Add mempool comparison tests
