+++ testutils/utils.go
@@ -4,8 +4,6 @@ import (
 	"math/rand"
 	"testing"
 
-	"cosmossdk.io/log"
-	"cosmossdk.io/math"
 	storetypes "cosmossdk.io/store/types"
 	txsigning "cosmossdk.io/x/tx/signing"
 	"github.com/cosmos/cosmos-sdk/client"
@@ -25,13 +23,6 @@ import (
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 	"github.com/cosmos/gogoproto/proto"
 
-	signerextraction "github.com/skip-mev/block-sdk/adapters/signer_extraction_adapter"
-	"github.com/skip-mev/block-sdk/block"
-	"github.com/skip-mev/block-sdk/block/base"
-	"github.com/skip-mev/block-sdk/block/mocks"
-	defaultlane "github.com/skip-mev/block-sdk/lanes/base"
-	"github.com/skip-mev/block-sdk/lanes/free"
-	"github.com/skip-mev/block-sdk/lanes/mev"
 	auctiontypes "github.com/skip-mev/block-sdk/x/auction/types"
 )
 
@@ -53,53 +44,6 @@ func CreateBaseSDKContext(t *testing.T) sdk.Context {
 	)
 
 	return testCtx.Ctx
-}
-
-func CreateMempool() *block.LanedMempool {
-	encodingConfig := CreateTestEncodingConfig()
-	signerExtractor := signerextraction.NewDefaultAdapter()
-
-	mevConfig := base.LaneConfig{
-		SignerExtractor: signerExtractor,
-		Logger:          log.NewNopLogger(),
-		TxEncoder:       encodingConfig.TxConfig.TxEncoder(),
-		TxDecoder:       encodingConfig.TxConfig.TxDecoder(),
-		AnteHandler:     nil,
-		MaxBlockSpace:   math.LegacyMustNewDecFromStr("0.3"),
-		MaxTxs:          0, // unlimited
-	}
-	factory := mev.NewDefaultAuctionFactory(encodingConfig.TxConfig.TxDecoder(), signerExtractor)
-	mevLane := mev.NewMEVLane(mevConfig, factory, factory.MatchHandler())
-
-	freeConfig := base.LaneConfig{
-		SignerExtractor: signerExtractor,
-		Logger:          log.NewNopLogger(),
-		TxEncoder:       encodingConfig.TxConfig.TxEncoder(),
-		TxDecoder:       encodingConfig.TxConfig.TxDecoder(),
-		AnteHandler:     nil,
-		MaxBlockSpace:   math.LegacyMustNewDecFromStr("0.3"),
-		MaxTxs:          0, // unlimited
-	}
-	freeLane := free.NewFreeLane[string](freeConfig, base.DefaultTxPriority(), free.DefaultMatchHandler())
-
-	defaultConfig := base.LaneConfig{
-		SignerExtractor: signerExtractor,
-		Logger:          log.NewNopLogger(),
-		TxEncoder:       encodingConfig.TxConfig.TxEncoder(),
-		TxDecoder:       encodingConfig.TxConfig.TxDecoder(),
-		AnteHandler:     nil,
-		MaxBlockSpace:   math.LegacyZeroDec(),
-		MaxTxs:          0, // unlimited
-	}
-	defaultLane := defaultlane.NewDefaultLane(defaultConfig, base.DefaultMatchHandler())
-
-	lanes := []block.Lane{mevLane, freeLane, defaultLane}
-	mempool, err := block.NewLanedMempool(log.NewNopLogger(), lanes, mocks.MockLaneFetcher{})
-	if err != nil {
-		panic(err)
-	}
-
-	return mempool
 }
 
 func CreateTestEncodingConfig() EncodingConfig {
